requirements.txt
stable-baselines3[extra]==1.0
pytest==6.2.2
pybullet==3.1.0
torchvision==0.9.0